Lift dichotomies state up to home page component.
