+++ mbti-type-builder/src/components/TypeAggregate.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+const TypeAggregate = () => {
+  return <div>TypeAggregate</div>;
+};
+
+export default TypeAggregate;
